backend/requirements.txt
# Flask core
Flask>=3.0.0
Werkzeug>=3.0.0
flask-cors>=4.0.0

# Database & ORM
Flask-SQLAlchemy>=3.1.0
SQLAlchemy>=2.0.0

# Serialization / validation
Flask-Marshmallow>=1.2.0
marshmallow>=3.20.0
marshmallow-sqlalchemy>=1.0.0

# Authentication
Flask-JWT-Extended>=4.6.0

# API documentation (Swagger UI)
flasgger>=0.9.7

# Configuration
python-dotenv>=1.0.0

# Real-time (WebSocket)
flask-socketio>=5.3.0
python-socketio>=5.10.0
